pyproject.toml
[tool.poetry]
name = "hooper-api"
version = "0.1.0"
description = ""
authors = ["Roddy Brown <roddy@notroddy.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
